Module ductgraph::graph::edgemap

Reexports

pub use self::edgematrix::VecMatrix;

Modules

edgematrix
A fixed sized array of edges used to implement Abstract graph with a fixed number of vertices.

Structs

EdgeTuple

MapGraph

MapGraphEntry

Neighbors

Traits

ToUsize
A conversion trait to usize