Module ductgraph::interface

Collection of traits providing operations on a generic graph data structure

Structs

Directed
Marker for directed edges : an edge from A to B is not the same as an edge from B to A.

Undirected
Marker for undirected edges : an edge between A and B is the same as an edge between B and A.

Traits

ConcreteGraphLike
Immutable operations on graphs.

Direction
Marker trait implemented on both edge direction markers : Directed and Undirected

Edge
Immutable operations on any edge starting from a vertex.

EdgeEntry
Unstable operations on any edge starting from a vertex.

EdgeMut
Mutable operations on any edge starting from a vertex.

Edges
Constructs an iterator on the edges starting from a vertex.

EdgesMut
Constructs an iterator on the edges starting from a vertex, yielding mutable references to the edge label.

Graph
A graph tructure.

GraphLike
A Graph structure with immutable operations on the vertices.

GraphLikeEntry
A Graph structure with unstable structural operations on the vertices.

GraphLikeMut
A Graph structure with mutable operations on the vertices.

Vertex
Immutable operations on a vertex entry.

VertexCreate
Vertex construction operation.

VertexEntry
Unstable operations on a vertex entry.

VertexMut
Mutable operations on a vertex entry.