Trait ductgraph::interface::ConcreteGraphLike

pub trait ConcreteGraphLike<V> {
    fn contains(&self, v: &V) -> bool;
}

Immutable operations on graphs.

Required Methods

fn contains(&self, v: &V) -> bool

Returns true if the graph contains the specified vertex

For concrete graphs, this method will always return true