Trait ductgraph::interface::Edge

pub trait Edge<'a, Q>: Sized {
    type V;
    type Label;
    type LabelGuard: Deref<Target=Self::Label> + 'a;

    fn edge_label(self, to: &'a Q) -> Option<Self::LabelGuard>;

    fn contains(self, to: &'a Q) -> bool { ... }
}

Immutable operations on any edge starting from a vertex.

Associated Types

type V

type Label

The type of the label of the edges.

type LabelGuard: Deref<Target=Self::Label> + 'a

Potentially a RAII guard which deref to a reference to a label.

Required Methods

fn edge_label(self, to: &'a Q) -> Option<Self::LabelGuard>

Returns a reference to the label of the edge starting from the current vertex to the given destination vertex.

If the edge doesn't exists, None is returned.

Provided Methods

fn contains(self, to: &'a Q) -> bool

Returns true if the graph contains an edge from the current vertex to the specified destination.