Trait ductgraph::interface::EdgeEntry

pub trait EdgeEntry<'a, Q> {
    type V;
    type Label;

    fn link(self, label: Self::Label, to: &'a Q) -> Option<Self::Label>;
    fn unlink(self, to: &'a Q) -> Option<Self::Label>;
}

Unstable operations on any edge starting from a vertex.

Associated Types

type V

type Label

The type of the label of the edges.

Required Methods

fn link(self, label: Self::Label, to: &'a Q) -> Option<Self::Label>

Adds an edge from the current vertex to the specified destination vertex with the given label.

If an edge between those two vertices already exists in the graph, the label is updated and the old label is returned. Else, None is returned.

fn unlink(self, to: &'a Q) -> Option<Self::Label>

Removes an edge from the current vertex to the specified destination vertex.

If no such edge exists, the method does nothing and returns None. If an edge exists, it is removed and its label is returned.