Trait ductgraph::interface::EdgesMut

pub trait EdgesMut<'a> {
    type Label;
    type V;
    type Edges: Iterator<Item=(&'a Self::V, &'a mut Self::Label)>;

    fn edges_mut(self) -> Self::Edges;
}

Constructs an iterator on the edges starting from a vertex, yielding mutable references to the edge label.

Associated Types

type Label

The type of edges label

type V

The type of vertex references

type Edges: Iterator<Item=(&'a Self::V, &'a mut Self::Label)>

The type of the iterator on the edges starting from the vertex

Required Methods

fn edges_mut(self) -> Self::Edges

Returns an iterator which yields a tuple (&'a Self::V, &'a mut Self::Label) for each edge starting from this vertex. The first element of the tuple is a reference to the destination vertex, and the second element of the tuple is a mutable reference to the label of the edge.