Trait ductgraph::interface::GraphLikeEntry

pub trait GraphLikeEntry<'a, V> {
    type VertexEntry;

    fn vertex_entry(&'a mut self, v: &'a V) -> Self::VertexEntry;
}

A Graph structure with unstable structural operations on the vertices.

Associated Types

type VertexEntry

The type of an unstable vertex.

Required Methods

fn vertex_entry(&'a mut self, v: &'a V) -> Self::VertexEntry

Retrieves an unstable vertex entry in the graph from a vertex reference.

Unstable operations are operation which modify the structure of the graph, for instance adding or removing an edge, removing a vertex, ...

Their is no notion of existence, because this interface is intended to be used both with abstract and concrete graphs.