Trait ductgraph::interface::GraphLikeMut

pub trait GraphLikeMut<'a, V> {
    type VertexMut;

    fn vertex_mut(&'a mut self, v: &'a V) -> Self::VertexMut;
}

A Graph structure with mutable operations on the vertices.

Associated Types

type VertexMut

The type of a mutable vertex.

Required Methods

fn vertex_mut(&'a mut self, v: &'a V) -> Self::VertexMut

Retrieves a mutable vertex entry in the graph from a vertex reference.

Their is no notion of existence, because this interface is intended to be used both with abstract and concrete graphs.