Trait ductgraph::interface::VertexCreate

pub trait VertexCreate {
    type Label;
    type V;

    fn create(self, label: Self::Label) -> Result<(Self::V, Option<Self::Label>), Self::Label>;
}

Vertex construction operation.

Associated Types

type Label

The type of a vertex label.

type V

The type of a vertex reference.

Required Methods

fn create(self, label: Self::Label) -> Result<(Self::V, Option<Self::Label>), Self::Label>

Add a new vertex in the graph with the specified label.

If the addition succeeded, a reference to the newly added vertex is returned. If this interface is used on a vertex entry, a vertex with this reference might already exists in the graph. In this case, the label is updated and the old label is returned.

If the addition fails, the given label is returned back.