Trait ductgraph::interface::VertexEntry

pub trait VertexEntry {
    type Label;

    fn remove(self) -> Option<Self::Label>;
}

Unstable operations on a vertex entry.

Associated Types

type Label

The type of a vertex label.

Required Methods

fn remove(self) -> Option<Self::Label>

Removes the vertex from the graph.

If the vertex entry exists in the graph, the vertex is removed, every edge starting from and pointing to this vertex are removed, and the vertex label is returned.

If the vertex entry doesn't exist, None is returned.