Module ductgraph::map

A generic interface on Map datastructures.

Reexports

pub use self::vec_list::VecListMap;

pub use self::stable_vec_map::StableVecMap;

Modules

stable_vec_map
A stable Map from unsigned integer keys to any value implemented as a vector. Search is in constant time

vec_list
A simple Map implemented as a list of key-value tuples in a Vec Provide linear time search, constant time removal and inserstion

Structs

BTreeMap
A map based on a B-Tree.

HashMap
A hash map implementation which uses linear probing with Robin Hood bucket stealing.

IterOrder
An iterator on an orderalence class defined by external order between Q and the key

Enums

ExtOrdering

Traits

Entry
A view into a single key-value pair placeholder in the map, which might already exists or be empty.

ExternalOrd

FixedMap
A mapping from Keys to Values with lookup functionnality for any type Q orderalent to the Key type

FixedSizedMap
A mapping from Keys to Values which can be extended with new mappings, but have a predefined size.

GrowableMap
A mapping from Keys to values which can be extended with new mappings but has no predefined size.

InternalIterator
In place filtering operation on map.

InternalMap
A Mapping from Keys to Values which are append only

InternalStableMap
A mapping from Keys to Values which can be extends with new mappings and from which values associated to a key can be retrived and modified. But the mapping is only exenstible with values. The key the new value is associated too is computed by the map itself.

IntoOrder
Convert a mapping into an iterator on a orderalence class of the key-value pairs.

Map
An unstable mapping from Keys to Values.

MapEntry
A mapping from Keys to Value which supports access and modification through the use of view into a single location in the map.

MapOwned
Any mapping from elements of type Key to elements of type Value

Occupied
A view into a single occupied location in a Map.

StableMap
A mapping from Keys to Values which can be extended with new mappings and from which values associated to a key can be retrieved and modified. The mapping is stable in the sense that no key-value pair can ever be removed.

UnstableFixedSizedMap
An unstable mapping from Keys to Values. Supports removal of key-value pairs, but the map has a fixed size.

Vacant
A view into a single empty location in a Map.

Implementations