Trait ductgraph::map::FixedMap

pub trait FixedMap<Q>: MapOwned where Self::Key: Borrow<Q> {
    fn get(&self, key: &Q) -> Option<&Self::Value>;
    fn get_mut(&mut self, key: &Q) -> Option<&mut Self::Value>;
    fn contains_key(&self, key: &Q) -> bool;
}

A mapping from Keys to Values with lookup functionnality for any type Q orderalent to the Key type

Required Methods

fn get(&self, key: &Q) -> Option<&Self::Value>

Returns a reference to the value corresponding to the key. The key may be any borrowed form of the map's key type, as long as the borrowed form properties (Eq, Hash, Ord, ...) match the properties on the keys type.

fn get_mut(&mut self, key: &Q) -> Option<&mut Self::Value>

Returns a mutable reference to the value corresponding to the key. The key may be any borrowed form of the map's key type, as long as the borrowed form properties (Eq, Hash, Ord, ...) match the properties on the keys type.

fn contains_key(&self, key: &Q) -> bool

Returns true if the map contains a value for the specified key.