Trait ductgraph::map::InternalMap

pub trait InternalMap: MapOwned {
    fn append(&mut self, value: Self::Value) -> Self::Key;
    fn is_empty(&self) -> bool;
}

A Mapping from Keys to Values which are append only

Required Methods

fn append(&mut self, value: Self::Value) -> Self::Key

Appends a new value to the map which will be associated with the next avaible key, and returns the key.

fn is_empty(&self) -> bool

Returns true if the map contains no key-value pair.