Trait ductgraph::map::InternalStableMap

pub trait InternalStableMap<Q>: InternalMap + FixedMap<Q> { }

A mapping from Keys to Values which can be extends with new mappings and from which values associated to a key can be retrived and modified. But the mapping is only exenstible with values. The key the new value is associated too is computed by the map itself.