Trait ductgraph::map::MapEntry

pub trait MapEntry<'a>: MapOwned {
    type Entry: Entry<'a, Value=Self::Value>;

    fn entry(&'a mut self, k: Self::Key) -> Self::Entry;
}

A mapping from Keys to Value which supports access and modification through the use of view into a single location in the map.

Associated Types

type Entry: Entry<'a, Value=Self::Value>

The type of a view into a single key-value pair placeholder

Required Methods

fn entry(&'a mut self, k: Self::Key) -> Self::Entry

Returns the placeholder in the map corresponding to the given key. The key doesn't need to exists in the map.