Trait ductgraph::map::Occupied

pub trait Occupied<'a> {
    type Value;

    fn insert(&mut self, value: Self::Value) -> Self::Value;
    fn remove(self) -> Self::Value;
    fn get(&mut self) -> &mut Self::Value;
    fn into_mut(self) -> &'a mut Self::Value;
}

A view into a single occupied location in a Map.

Associated Types

type Value

Required Methods

fn insert(&mut self, value: Self::Value) -> Self::Value

Replace the value of the key-value pair associated with this location with the given value, and returns the old value associated with this location.

fn remove(self) -> Self::Value

Takes the value out of the entry and returns it. The key-value pair is removed from the Map.

fn get(&mut self) -> &mut Self::Value

Returns a mutable reference to the value associated with this location.

fn into_mut(self) -> &'a mut Self::Value

Converts this entry into a mutable reference to the value associated with this location.