Trait ductgraph::map::UnstableFixedSizedMap

pub trait UnstableFixedSizedMap<Q>: FixedMap<Q> + FixedSizedMap {
    fn remove(&mut self, key: &Q) -> Option<Self::Value>;
}

An unstable mapping from Keys to Values. Supports removal of key-value pairs, but the map has a fixed size.

Required Methods

fn remove(&mut self, key: &Q) -> Option<Self::Value>

Remove a key-value pair from the map.

If a value is present in the map for the specified key, this value is removed from the map and returned. If no such value exists, None is returned. The key may be any borrowed form of the map's key type, as long as the borrowed form properties (Eq, Hash, Ord, ...) match the properties on the keys type.