Struct ductgraph::map::vec_list::VecListMap

pub struct VecListMap<K, V> {
    // some fields omitted
}

A simple Map implemented as a list of key-value pair in a Vec Provide linear time search, constant time removal and inserstion

Methods

impl<K, V> VecListMap<K, V>

fn new() -> VecListMap<K, V>

Construct a new empty map with default capacity

fn with_capacity(cap: usize) -> VecListMap<K, V>

Constructs a new, empty map with the specified capacity.

The map will be able to hold exactly capacity elements without reallocating. If capacity is 0, the map will not allocate.