Module ductgraph::uuid

An interface to manage construction of unique clonable handles

A graph structure might be cyclic, and has no clear notion of ownership. Nonetheless, some kind of reference is needed to implement the graph. This module provides an interface to create generic name handles to act as references to vertices in a graph structure. This interface is used only when the handles are generated by the graph structure when a new vertex is created. The users can also build a graph structure which doesn't manage the handles internally, and provide the handle at the time of vertex creation.

Structs

Factory
A handle factory for a graph structure internally managing the handles.

InternalFactory
A handle factory for a graph structure internally managing the handles, but not using an user provided factory. For this kind of graph, handles are usually integer and are managed by the graph code itself.

UnitFactory
The Unit factory. The graph doesn't manage the vertex handles itself, and ask for a handle to the user at the vertex creation.

Traits

StableUUIDFactory
A vertex handle factory with a state able to check if a handle is valid (ie has already been returned by the factory.

UUIDFactory
A vertex handle factory.